Introduce the Installation Type enum and property
This commit adds the Type enum & property to the Installation object,
for the purpose of controlling which files to install.

------------------------------------------------------------------------

TYPES
-----

NORMAL

Install the os library and in-game configuration gui - the most common
install people would need.

MINIMAL

Install the os library (used for portability; only the hce folder needs
to be accessed).

COMPLETE

Installs the os library, in-game gui, hek modules (access is needed to
system and program data).